// src/app/modules/hcm/shared/models/masters/company-jobposition.ts
export class companyjobposition {
  id: number = 0;
  company: number = 0;
  hierarchicalLevel: number = 0;
  mainJobPosition: number = 0;
  mtJobPosition: number = 0;
  fixedPositions: number = 0;
  temporaryPositions: number = 0;
  name: string = "";
  description: string = "";
  duties: string = "";
  goals: string = "";
  riskAnalysis: string = "";
  supervisoryPosition: boolean = false;
}